+++ stdlib/REPL/test/TerminalMenus/quit_selection.jl
@@ -5,13 +5,13 @@
     @test simulate_input(menu, 'q') == -1
 
     menu = MultiSelectMenu(["a", "b", "c"]; warn=false)
-    @test simulate_input(menu, 'q') == Set{Int64}()
+    @test simulate_input(menu, 'q') == Set{Int}()
 
     menu = RadioMenu(["a", "b", "c"]; charset=:ascii)
     @test simulate_input(menu, 'q') == -1
 
     menu = MultiSelectMenu(["a", "b", "c"]; charset=:ascii)
-    @test simulate_input(menu, 'q') == Set{Int64}()
+    @test simulate_input(menu, 'q') == Set{Int}()
 
     menu = RadioMenu(["a", "b", "c"]; on_cancel=nothing)
     @test simulate_input(menu, 'q') |> isnothing
